requirements.txt
pandas>=1.3.0
requests>=2.26.0
google-auth>=2.3.3
google-auth-oauthlib>=0.4.6
google-auth-httplib2>=0.1.0
google-api-python-client>=2.0.0
gspread>=5.0.0
python-dotenv>=0.19.0
flask>=2.0.0 